Mark e2e tests as passed after undeploy passes
This will prevent changes that break the helm uninstall operations to
get merged. This is important now that we have a helm distribution that
we ship to customers.

+++ pipelines/scripts/undeploy-korifi.sh
@@ -21,7 +21,9 @@ undeploy_cf() {
 
   kubectl delete namespace cf --ignore-not-found
 
-  helm delete --namespace korifi korifi --wait
+  if helm status --namespace korifi korifi; then
+    helm delete --namespace korifi korifi --wait
+  fi
 }
 
 main() {
